Rename main to old_main to indicate it is the original file
